added JSDocs to many functions. refactor listingsRender function
