+++ src/js/api/constants.js
@@ -1,4 +1,4 @@
-//VARIABLES FOR AUTHENTICATION
+//VARIABLES
 
 export const API_KEY = '4ac6445f-6305-4510-89a4-cb5df85ce9c1';
