+++ src/js/api/profiles/profileBids.js
@@ -4,6 +4,23 @@ import { authFetch } from '../fetch/authFetch';
 
 const action = '/profiles';
 
+/**
+ * Retrieves all bids made by the current user's profile, including the associated listings.
+ *
+ * This function loads the user's profile from storage, fetches their bid history from the API,
+ * and includes the listings that the bids are related to.
+ *
+ * @async
+ * @function profileBids
+ * @returns {Promise<Object>} The profile's bid history with associated listings.
+ * @throws {Error} If the profile name is not available or the request fails.
+ *
+ * @example
+ * profileBids()
+ *   .then(bids => console.log(bids))
+ *   .catch(error => console.error(error));
+ */
+
 export async function profileBids() {
   const profile = load('profile');
   const name = profile.name;
@@ -13,10 +30,8 @@ export async function profileBids() {
   }
 
   const url = `${API_BASE}${API_AUCTIONS}${action}/${name}/bids?_listings=true`;
-  // console.log(url);
 
   const response = await authFetch(url, 'GET');
-  // console.log('profileBids response:', response);
 
   return response;
 }
