+++ src/js/render/helpers/create/listingsCard/index.js
@@ -1,6 +1,21 @@
+/**
+ * Re-exports modules from various component files for easier access.
+ *
+ * This file re-exports all the functions from multiple files in the same directory,
+ * including `cardContainer.js`, `bodyContainer.js`, `footerContainer.js`, `mediaContainer.js`,
+ * and `viewButtonContainer.js`. This allows other modules to import any of these exports from
+ * a single file, improving the organization of module imports.
+ *
+ * @module components
+ * @exports * from './cardContainer.js'
+ * @exports * from './bodyContainer.js'
+ * @exports * from './footerContainer.js'
+ * @exports * from './mediaContainer.js'
+ * @exports * from './viewButtonContainer.js'
+ */
+
 export * from './cardContainer.js';
 export * from './bodyContainer.js';
 export * from './footerContainer.js';
-//export * from './mediaCarousel.js';
 export * from './mediaContainer.js';
 export * from './viewButtonContainer.js';
