+++ src/js/render/helpers/create/listingsCard/viewButtonContainer.js
@@ -1,3 +1,19 @@
+/**
+ * Creates and returns a button element that navigates to the listing details page when clicked.
+ *
+ * This function generates a styled "View" button that redirects the user to the listing page
+ * based on the provided `postData` ID when clicked.
+ *
+ * @function viewButtonContainer
+ * @param {Object} postData - The data object containing the listing information.
+ * @param {string} postData.id - The unique ID of the listing to view.
+ * @returns {HTMLButtonElement} The created button element.
+ *
+ * @example
+ * const viewButton = viewButtonContainer(postData);
+ * document.body.appendChild(viewButton);
+ */
+
 export function viewButtonContainer(postData) {
   const viewButtonContainer = document.createElement('button');
