+++ src/js/render/helpers/create/specificCard/index.js
@@ -1,3 +1,17 @@
+/**
+ * Re-exports modules from specific container files for easier access.
+ *
+ * This file re-exports all the functions from multiple specific container files, including:
+ * `specificCardContainer.js`, `specificBodyContainer.js`, `specificFooterContainer.js`,
+ * and `specificMediaContainer.js`. This allows importing them from a single file.
+ *
+ * @module specificContainers
+ * @exports * from './specificCardContainer.js'
+ * @exports * from './specificBodyContainer.js'
+ * @exports * from './specificFooterContainer.js'
+ * @exports * from './specificMediaContainer.js'
+ */
+
 export * from './specificCardContainer.js';
 export * from './specificBodyContainer.js';
 export * from './specificFooterContainer.js';
