+++ src/js/render/helpers/create/specificCard/placeBid.js
@@ -1,3 +1,17 @@
+/**
+ * Returns the container for placing a bid.
+ *
+ * This function is a placeholder that returns the `placeBidContainer`, which likely handles the UI
+ * or logic for placing a bid.
+ *
+ * @function placeBid
+ * @returns {HTMLElement} The container element responsible for the bid placement.
+ *
+ * @example
+ * const bidContainer = placeBid();
+ * document.body.appendChild(bidContainer);
+ */
+
 export function placeBid() {
   return placeBidContainer;
 }
