+++ src/js/render/helpers/create/specificCard/specificBodyContainer.js
@@ -1,3 +1,18 @@
+/**
+ * Creates and returns a styled `div` element as a specific card body container.
+ *
+ * This function generates a `div` element with classes for a responsive layout,
+ * including flexbox configurations that align content vertically and horizontally.
+ * It is designed for use as the body of a card component.
+ *
+ * @function specificbodyContainer
+ * @returns {HTMLDivElement} The created card body container element.
+ *
+ * @example
+ * const cardBody = specificbodyContainer();
+ * document.body.appendChild(cardBody);
+ */
+
 export function specificbodyContainer() {
   const cardBody = document.createElement('div');
