+++ src/js/render/index.js
@@ -1,3 +1,18 @@
+/**
+ * Re-exports rendering functions for various components.
+ *
+ * This file re-exports functions for rendering listings, profiles, and carousels, making them accessible
+ * from a single module. The functions include rendering limited listings, individual listings,
+ * all listings, profiles, and carousels.
+ *
+ * @module renderFunctions
+ * @exports * from './limitListingsRender.js'
+ * @exports * from './listingRender.js'
+ * @exports * from './listingsRender.js'
+ * @exports * from './renderProfile.js'
+ * @exports * from './renderCarousel.js'
+ */
+
 export * from './limitListingsRender.js';
 export * from './listingRender.js';
 export * from './listingsRender.js';
