+++ src/js/ui/listeners/isBottomOfPage.js
@@ -1,0 +1,3 @@
+export function isBottomOfPage() {
+  return window.innerHeight + window.scrollY >= document.body.offsetHeight - 100;
+}
