+++ src/js/ui/listeners/scrollListener.js
@@ -1,0 +1,8 @@
+export function scrollListener() {
+  // Add event listener for scroll
+  window.addEventListener('scroll', () => {
+    if (isBottomOfPage() && !isLoading) {
+      listingsRender();
+    }
+  });
+}
